main/agents/agent-pptx-copy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-generation-agent"
version = "2.0.0"
description = "A2A-compatible PowerPoint Generation Agent (v2 — with images & charts)"
requires-python = ">=3.10"
dependencies = [
    "a2a-sdk>=0.3.0",
    "click>=8.1.8",
    "openai>=1.57.0",
    "pydantic>=2.11.4",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "python-pptx>=0.6.23",
    "lxml>=4.9.0",
    # v2 additions
    "matplotlib>=3.8.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["src"]
